# perf: cache parsed requirements