--- a/src/poetry/core/version/requirements.py
+++ b/src/poetry/core/version/requirements.py
@@ -1,12 +1,19 @@
 from __future__ import annotations
 
+import functools
 import urllib.parse as urlparse
+
+from typing import TYPE_CHECKING
 
 from poetry.core.constraints.version import parse_constraint
 from poetry.core.constraints.version.exceptions import ParseConstraintError
 from poetry.core.version.grammars import GRAMMAR_PEP_508_CONSTRAINTS
 from poetry.core.version.markers import _compact_markers
 from poetry.core.version.parser import Parser
+
+
+if TYPE_CHECKING:
+    from collections.abc import Sequence
 
 
 class InvalidRequirementError(ValueError):
@@ -61,7 +68,9 @@ class Requirement:
         else:
             self.url = None
 
-        self.extras = [e.value for e in parsed.scan_values(lambda t: t.type == "EXTRA")]
+        self.extras: Sequence[str] = [
+            e.value for e in parsed.scan_values(lambda t: t.type == "EXTRA")
+        ]
         constraint = next(parsed.find_data("version_specification"), None)
         constraint = ",".join(constraint.children) if constraint else "*"
 
@@ -102,3 +111,8 @@ class Requirement:
 
     def __repr__(self) -> str:
         return f"<Requirement({str(self)!r})>"
+
+
+@functools.cache
+def parse_requirement(requirement_string: str) -> Requirement:
+    return Requirement(requirement_string)