<!-- This is just a reminder about the most common mistakes. Please make sure that you tick all *appropriate* boxes. But please read our [contribution guide](https://python-poetry.org/docs/contributing/) at least once, it will save you unnecessary review cycles! -->

- [ ] Added **tests** for changed code.
- [ ] Updated **documentation** for changed code.

Caching the requirements should be safe because we use them read-only.

Testing with the `pyproject.toml` from https://github.com/python-poetry/poetry/issues/9956#issuecomment-2582451109:

poetry version|lock (--regenerate)|lock (--no-update)
---|---|---
main|98 s|17.5 s
PR|73 s|11.5 s

## Summary by Sourcery

Cache parsed requirement strings to improve performance, especially when regenerating lock files.

Enhancements:
- Cache requirement strings using `functools.cache`.

Tests:
- Added tests to verify the caching behavior and ensure there are no regressions.